requirements.txt
camel-ai
streamlit
python-dotenv
pdfplumber
groq
